new classes and solutions and test

+++ src/main/java/ar/com/problems/numbers/SumDigitalRoots.java
@@ -28,7 +28,7 @@ public class SumDigitalRoots {
                
 
                 for (Integer intDigit : arrydigits) {
-                    sumDigits += intDigit.intValue();
+                    sumDigits += intDigit;
                 }
                 arrydigits = new ArrayList<>();
